pyproject.toml
[project]
name = "board-api"
version = "0.1.0"
description = "GraphQL backend for the Board app built with FastAPI, Strawberry, PostgreSQL, and GitHub auth."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.5",
    "authlib>=1.6.11",
    "fastapi>=0.128.8",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "psycopg[binary]>=3.2.13",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.49",
    "strawberry-graphql>=0.283.3",
    "uvicorn>=0.39.0",
]

[dependency-groups]
dev = [
    "anyio>=4.12.1",
    "mypy>=1.19.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.15.12",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
check_untyped_defs = true
exclude = ["^tests/"]

# Progressive start to not brake flow from start
disallow_untyped_defs = false
disallow_incomplete_defs = false
